Move PhotoMarkers class to cyclestreets-view library; make use of new drawables
